replace ectool with framework_tool
remove ectool in uninstall legacy process
