--- fetch/ectool/linux/gitlab_job_id
@@ -1,1 +1,0 @@
-899
