--- fetch/ectool/linux/hash.sha256
@@ -1,1 +1,0 @@
-ab94a1e9a33f592d5482dbfd4f42ad351ef91227ee3b3707333c0107d7f2b1b0
