+++ fetch/framework_tool/linux/release.version
@@ -1,0 +1,1 @@
+v0.6.4
