+++ fetch/framework_tool/linux/target.url
@@ -1,0 +1,1 @@
+https://github.com/FrameworkComputer/framework-system/releases/download/%RELEASE_VERSION%/framework_tool
